third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-xhr-in-main-frame.html

<html>
<body>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
    testRunner.setCanOpenWindows();
    testRunner.setCloseRemainingWindowsWhenComplete(true);
    testRunner.overridePreference("WebKitAllowRunningInsecureContent", true);
}
if (window.internals)
    internals.settings.setLaxMixedContentCheckingEnabled(false);


window.addEventListener("message", function (e) {
    if (window.testRunner)
        testRunner.notifyDone();
}, false);

</script>
<p>This test opens a HTTPS window that loads insecure data via XHR.  We should
trigger a mixed content callback because the main frame in the window is HTTPS but
now has insecure data.</p>
<script>
window.open("https://127.0.0.1:8443/xmlhttprequest/access-control-response-with-body.html");
</script>
</body>
</html>
